requirements.txt
flask==2.2.2
openai==0.28.0
python-dotenv==1.0.0
waitress==2.1.2